pyproject.toml
[tool.poetry]
name = "sypro-coursework"
version = "0.1.0"
description = ""
authors = ["jenuk <doctor_who_1996@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
openpyxl = "^3.1.5"
load-dotenv = "^0.1.0"
env = "^0.1.0"
python-dotenv = "^1.0.1"
transactions = "^0.2.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
mypy = "^1.13.0"
isort = "^5.13.2"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[flake8]
max_line_length = 119

[tool.blake]
line-length = 119

[tool.isort]
line_length = 119


